refactor(create-plugin): delete update.migrate.command file
